Added some new pdfs which would be a good source of keywords. FYI!!! pdfminer_multiple will need some massaging since it does not always read the pdfs very well and throws some errors.
Added folder to pdfminer called "clean" normalize_text.py iterates thru
the pdfs and generates lemmas and stores them in corresponding text files
under the same name in this folder.

normalize_text.py goes through all converted texts and cleans them and
extracts lemmas along with their counts, the summed total of all these
lemmas across the pdfs is stored in "to_DB.txt" in principle, the higher
the count, the more likely this lemma can be used to suggest one is on-topic.

added getFocusScore.py which accesses mongoDB and finds the average focus
score for the collection in sessions

Signed-off-by: vorado <vorado@live.com>